+++ packages/vsce-api/__tests__/__unit__/ResourceTypes.test.ts
@@ -24,6 +24,7 @@ describe("ResourceType tests", () => {
       "CICSPipeline",
       "CICSWebService",
       "CICSJVMServer",
+      "CICSBundle",
     ]);
   });